slidechain: prevent double peg out (#46)
Adds a pre-export tx that creates a temporary account which is merged back into the exporter upon peg-out. If there is an accidental double peg-out, the transaction will fail since the temporary account will no longer exist, preventing funds from being sent from the custodian twice.

This doesn't prevent against a malicious attack that tries to recreate the temporary account after merge and uses BumpSequence to set it to the correct sequence number, and is also vulnerable to failure if the exporter closes their account between the pre-export tx and peg-out.