slidechain: add txvm programs for import uniqueness (#54)
Adds TxVM programs to guarantee uniqueness of pegged-in Stellar transactions and their corresponding imports on TxVM. The `createTokenProg` creates a uniqueness token and is run before submitting the peg-in transaction to the Stellar network. The `consumeTokenProg` consumes said token and thus ensures that the import for a specific peg-in can only happen once. The `importIssuanceProg` calls the `consumeTokenProg`'s contract to produce various arguments for the import transaction and is described in more detail [here](https://github.com/interstellar/slingshot/blob/main/slidechain/Pegging.md#importing). Note that this PR does not explicitly call or submit these transactions; including these changes into the greater import workflow is left for a future PR.

https://trello.com/c/SnZp7YD3, 1-3