slidechain: add double import protection (#52)
Adds protection against double import.

Adds RPCs to `slidechaind` to wait on submitted TxVM transactions to hit the chain, and record peg-ins. Uses RPCs to create and submit the uniqueness token to TxVM, and modifies the import transaction to consume said token before issuing the pegged-in asset. Re-keys the `pegs` table on a unique transaction hash, and changes the schema of `pegs` also to track whether a transaction has hit the Stellar network. Moves insertion from the `watch` goroutine to between submitting the uniqueness token and the Stellar peg-in transaction. Updates associated tests.

https://trello.com/c/SnZp7YD3, 4-9