slidechain: fix amount discrepancy bug in peg cmd (#141)
Fixes an amount discrepancy bug in `cmd/peg`. Lumen amounts were being passed as stroops to the pre-peg-in TxVM transaction, but as lumens to the import transaction. This discrepancy manifested itself as an `invalid prevout` error when submitting the import transaction, because the contract stack of the reconstructed pre-peg-in snapshot differed from that of the actual transaction.