slidechain: streamline peg-in RPC (#140)
Streamlines peg-in RPC and surrounding logic. Previously, the `record` RPC only recorded the peg-in transaction in the `pegs` table. As a result, it required a prior server round-trip to submit and wait on the pre-peg-in TxVM transaction. Now, the RPC (renamed to `prepegin`) also builds and submits the pre-peg-in transaction. All pre-peg-in related logic is moved into a single file.