slidechain: add peg-out smart contract (#83)
Implements a smart contract for peg-out. The better peg-out protocol (#49) suggests that the txvm-side export transaction lock the funds in a smart contract. If the peg-out Stellar transaction succeeds, that contract retires the funds; else, it repays them to the exporter. This PR accordingly updates the export transaction; implements the aforementioned post-export contract; adds a go-routine to watch for the Stellar transaction; and updates tests and associated binaries as needed.

https://trello.com/c/boButuVu