zkvm: remove nonces (#271)
This removes nonces from the design. This dramatically simplifies blockchain state and validation rules: every tx must spend an input to be unique. Initial anchoring can be provided in a number of ways: pegged assets from a parent chain, and/or minted synthetic nonce-units at each block, etc. See #230 and #249 for details. To bootstrap anchoring, the initial block can start with an initial UTXO set (this API is necessary to have until we have minteable nonce-units (if at all), and even then it may be useful in its own right).

UTXO set commitment is changed to a normal merkle tree (instead of radix/patricia trie), in order to support utreexo (see #258). The state machine spec is still assuming presence of the entire set, this will be updated later.

Closes #262.
Closes #263.