zkvm: non-copyable Program, Expression and Constraint (#304)
This adds a non-copyable type `Program` which is separate from `Data` type. Instructions `call` and `delegate` expect Program type, instruction `dup` fails if used with a Program.

Also, the `Expression` and `Constraint` are made non-copyable and non-droppable. Variables are left to be copyable, so they can participate in constraints and then be used in the construction of Values.

Rationale is covered in #301 & #306.

Closes #301
Closes #306