accounts: new crate for Account and payment protocol (#327)
Account is a private abstraction over a collection of UTXOs. Each node maintains one or more accounts identified by a root xpub/xprv keypair (see `Keytree` crate) and a current _sequence number_. 

When account is requesting a payment, it increments the sequence number, derives the one-time key and unique blinding factors, wraps them in a `Receiver` object and passes it to the sender. Sender then composes a transaction, selecting some existing utxos to spend, and creating new utxo to return leftover funds as _change_.

After tx is committed to the blockchain, both sender and recipient receive the new block, detect payment and update their state accordingly.

- [x] Accounts and Receivers
- [x] Simple UTXO selection algorithm
- [x] Full workflow test with Alice and Bob
- [x] Various tweaks to the existing APIs to make it all work

Closes #322