zkvm: allow dropping Program/Expression/Constraint (#407)
Fixes https://github.com/stellar/slingshot/issues/406

The Program/Expression/Constraint are safe to drop types, but not safe to copy due to DoS risk. This was not implemented since we used a single category of "copyable types" for both copying and dropping. This patch introduces another category of "droppable types".

Specifically, Expression should be allowed to be dropped because of the `range` instruction that leaves it on stack. If one needs an inequality check, the left-on-stack expression must be dropped to clear the stack. In turn, `range` cannot consume the expression because then it cannot be used in downstream expressions, and we cannot allow copying for expressions for the DoS reasons.

Constraint and Program are also permitted to be dropped as there is no reason to prohibit that.