demo: fix rendering of account (node/show template) (#477)
### What
Change where the account sequence and xpub are retrieved from on the
node/show page.

### Why
The account package was refactored in
dffffb2fe0bd85467479e9e8422d63cc973e9a14 and that resulted in the
sequence moving from the account object to the wallet object, and the
xpub is no longer stored in an object that has been serialized.

Updating where the sequence should come from is trivial and not
controversial.

Updating where the xpub could be done a couple of ways. We could expand
the JSON serialized wallet to include it, but that requires changing the
account package and the xpub would only be added to the struct for the
demo, so instead I added it to the context for rendering explicitly. I
could see reasons to go both ways though.